# server/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": " This is the server is of a Pokedex application. All the information about the pokemon comes from the official API of pokemon https://pokeapi.co/ .",
      "version": "1.0.0",
      "title": "Pokedex",
      "contact": {
        "email": "danielurrutxua@opendeusto.es"
      }
    },
    "host": "localhost:5000",
    "basePath": "/",
...
    "schemes": [
      "https",
      "http"
    ],
    "paths": {
      "/pokemon": {
        "get": {
          "tags": [
            "pokemon"
          ],
          "summary": "Get all pokemon",
          "description": "Returns all pokemon existing in the PokeApi",
          "operationId": "getAllPokemon",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "username",
              "description": "username to send also the favourite pokemon",
              "required": false,          
              "type":"string"
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/PokemonList"
                }
              }
            },
            "400": {
              "description": "Error connecting to PokeApi"
            }
          },
          "x-swagger-router-controller": "Pokemon"
        }
      },
      "/pokemon/{pokemonId}": {
        "get": {
          "tags": [
            "pokemon"
          ],
          "summary": "Find pokemon by ID",
          "description": "Returns a single pokemon",
          "operationId": "getPokemonById",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "pokemonId",
              "in": "path",
              "description": "ID of pokemon to return",
              "required": true,
              "type": "integer",
              "format": "int64"
              
            },
            {
              "in": "body",
              "name": "username",
              "description": "username to send also the favourite value",
              "required": false,          
              "type":"string"
            }
          ],
          "security": [
            {
              "api_key": []
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "schema": {
                "$ref": "#/definitions/PokemonDetailed"
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Pokemon not found"
            }
          },
          "x-swagger-router-controller": "Pokemon"
        }
      },
...
    },
    "definitions": {
...
      "PokemonList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "example": "pikachu"
          },
          "types": {
            "type": "array",
            "example": [
              "water",
              "fire"
            ],
            "items": {
              "type": "string"
            }
          },
          "favourite": {
            "type": "boolean",
            "example": true
          }
        },
        "xml": {
          "name": "Pokemon"
        },
        "example": {
          "types": [
            "water",
            "fire"
          ],
          "name": "pikachu",
          "id": 0,
          "favourite": true
        }
      },
      "PokemonDetailed": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "height": {
            "type": "integer"
          },
          "weight": {
            "type": "integer"
          },
          "types": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "abilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "stats": {
            "type": "object",
            "properties": {
              "hp":{
                "type":"integer"
              },
              "attack": {
                "type":"integer"
              },
              "defense":{
                "type":"integer"
              },
              "special-attack":{
                "type":"integer"
              },
              "speed": {
                "type":"integer"
              },
              "special-defense":{
                "type":"integer"
              }
            },
            "items": {
              "type": "integer",
              "format": "int64"
            }
          },
          "favourite": {
            "type": "boolean",
            "example": true
          }
        },
        "example": {
          "name": "pikachu",
          "id": 0,
          "weight": 120,
          "height": 243,
          "types": [
            "water",
            "fire"
          ],
          "abilities": [
            "ability1",
            "ability2"
          ],
          "stats": {
            "hp": 48,
            "attack": 48,
            "defense": 48,
            "special-attack": 48,
            "speed": 48,
            "special-defense": 48
          },
          "favourite": true
        }
      },
...
    }
  }